requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
Pygments==2.19.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
shellingham==1.5.4
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
